+++ intent-generation/src/main/java/com/multiassetoms/intentgeneration/model/OrderIntentNotFoundException.java
@@ -1,0 +1,8 @@
+package com.multiassetoms.intentgeneration.model;
+
+public class OrderIntentNotFoundException extends RuntimeException {
+
+    public OrderIntentNotFoundException(String message) {
+        super(message);
+    }
+}